Fixed failing tests and updated folders.
