+++ EmbeddedSocial/Sources/Common/AsyncOperation.swift
